Fragment transitions!
Includes a fix to NavigationFragment that feels like a hack: unsetting
the Adapter in onDestroyView() to prevent a crash that occurs when
closing the Fragent. From what I can gather, the crash occurs because the
special Adapter wrapper that the framework adds when using ListView
headers is looking for the header, but it's been disposed. Strangely, the
error only occurs when the transition is used...