Fixed multi-select playback track ordering
Instead of building a query that uses a WHERE IN clause - a query whose
ordering cannot be controlled - use multiple successive queries to ensure
that tracks are added to the playlist in the order that they were
selected.

This changed caused a crash to occur when attempting to add multiple
items, because the PlaybackService's Cursor would be closed and stale
when the MediaSessionManager tried to use it to update the metadata for
the curret track. PlayingItem now retrieves the data from the Cursor and
stores it in instance variables, preventing this crash.