PlaybackService takes Lists, not Cursors
PlaybackService will run its own queries to get Cursors based on the
Lists it is supplied. This fixes issues with MainActivity's CursorLoader
resetting the Cursor passed to the PlaybackService before its had a chance
to add the tracks.

This change also removes the stutter that was seen after clicking one of
the multi-select action buttons. Huzzah!